PEP8 formating issues ok. Users can only edit/delete items if they created it.

+++ project.py
@@ -1,8 +1,7 @@
-from flask import Flask, render_template, request, redirect, jsonify, url_for, flash
 from sqlalchemy import create_engine, asc, func
-from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, Item, User
 from flask import session as login_session
+from database_setup import Base, Category, Item, User
+from sqlalchemy.orm import sessionmaker
 import random
 import string
 from oauth2client.client import flow_from_clientsecrets
@@ -11,6 +10,8 @@ import httplib2
 import json
 from flask import make_response
 import requests
+from flask import Flask, render_template, request, redirect
+from flask import jsonify, url_for
 
 app = Flask(__name__)
 
@@ -35,8 +36,8 @@ def showLogin():
     login_session['state'] = state
     # return "The current session state is %s" % login_session['state']
     return render_template('login.html', STATE=state)
-#
-#
+
+
 @app.route('/gconnect', methods=['POST'])
 def gconnect():
     # Validate state token
@@ -89,7 +90,8 @@ def gconnect():
     stored_access_token = login_session.get('access_token')
     stored_gplus_id = login_session.get('gplus_id')
     if stored_access_token is not None and gplus_id == stored_gplus_id:
-        response = make_response(json.dumps('Current user is already connected.'), 200)
+        response = make_response(json.dumps('Current user is \
+            already connected.'), 200)
         response.headers['Content-Type'] = 'application/json'
         return response
 
@@ -121,24 +123,28 @@ def gconnect():
     output += '!</h1>'
     output += '<img src="'
     output += login_session['picture']
-    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+    output += ' " style = "width: 300px; height: 300px;border-radius: \
+        150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
     flash("you are now logged in as %s" % login_session['username'])
     print "done!"
     return output
-#
-#     # DISCONNECT - Revoke a current user's token and reset their login_session
+
+
+# DISCONNECT - Revoke a current user's token and reset their login_session
 @app.route('/gdisconnect')
 def gdisconnect():
     access_token = login_session.get('access_token')
     if access_token is None:
         print 'Access Token is None'
-        response = make_response(json.dumps('Current user not connected.'), 401)
+        response = make_response(json.dumps('Current user \
+                not connected.'), 401)
         response.headers['Content-Type'] = 'application/json'
         return response
     print 'In gdisconnect access token is %s', access_token
     print 'User name is: '
     print login_session['username']
-    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['access_token']
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % \
+        login_session['access_token']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print 'result is '
@@ -153,10 +159,12 @@ def gdisconnect():
         response.headers['Content-Type'] = 'application/json'
         return response
     else:
-        response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        response = make_response(json.dumps('Failed to revoke token for \
+                given user.', 400))
         response.headers['Content-Type'] = 'application/json'
         return response
-#
+
+
 # #disconnect from any provider, still only google
 @app.route('/disconnect')
 def disconnect():
@@ -173,14 +181,16 @@ def disconnect():
         flash("You were not logged in to begin with.")
         return redirect(url_for('showCatalog'))
 
-#All categories JSON
+
+# All categories JSON
 @app.route('/catalog/JSON')
 def catalogJSON():
     catalog_all = session.query(Category).all()
     items_all = session.query(Item).all()
-    catalog = {"Category": [cat.serialize for cat in catalog_all]}
+    catalog = {"Category": [cate.serialize for cate in catalog_all]}
     for cate in catalog["Category"]:
-        cate["Item"] = [item.serialize for item in items_all if item.category_id == cate['id']]
+        cate["Item"] = [item.serialize for item in items_all if
+                        item.category_id == cate['id']]
     return jsonify(catalog)
 
 
@@ -194,46 +204,67 @@ def showCatalog():
     else:
         return render_template('catalog.html', catalog=catalog)
 
-#show a specific Item
+
+# show a specific Item
 @app.route('/catalog/<string:cat_name>/items')
 def showItems(cat_name):
     catalog = session.query(Category).order_by(asc(Category.name))
     cat = session.query(Category).filter_by(name=cat_name).one()
-    items = session.query(Item).filter_by(category_id= cat.id).order_by(Item.name)
+    items = session.query(Item).filter_by(category_id=cat.id).\
+        order_by(Item.name)
     if 'username' not in login_session:
-        return render_template('publicmenu.html', items = items, catalog=catalog, cat=cat, count = get_count(items))
+        return render_template('publicmenu.html', items=items,
+                               catalog=catalog, cat=cat,
+                               count=get_count(items))
     else:
-        return render_template('menu.html', items = items, catalog=catalog, cat=cat, count = get_count(items))
+        return render_template('menu.html', items=items, catalog=catalog,
+                               cat=cat, count=get_count(items))
+
 
 @app.route('/catalog/<string:cat_name>/<string:item_name>')
 def showOneItem(cat_name, item_name):
     cat = session.query(Category).filter_by(name=cat_name).one()
-    item = session.query(Item).filter_by(category_id=cat.id, name=item_name).one()
+    item = session.query(Item).filter_by(category_id=cat.id,
+                                         name=item_name).one()
     if 'username' not in login_session:
-        return render_template('publicitem.html', item = item, category_id = cat.id)
+        return render_template('publicitem.html', item=item,
+                               category_id=cat.id)
     else:
-        return render_template('item.html', item = item, category_id = cat.id)
+        return render_template('item.html', item=item, category_id=cat.id)
 
-#Create new Item
+
+# Create new Item
 @app.route('/catalog/new/', methods=['GET', 'POST'])
 def newItem():
     catalog = session.query(Category).order_by(asc(Category.name))
     if request.method == 'POST':
-        selectedCategory = session.query(Category).filter_by(name = request.form['category']).one()
+        selectedCategory = session.query(Category).\
+                        filter_by(name=request.form['category']).one()
         if request.form['title']:
-            newItem = Item(name=request.form['title'], description=request.form['description'], category_id=selectedCategory.id)
+            print login_session['user_id']
+            newItem = Item(name=request.form['title'],
+                           description=request.form['description'],
+                           category_id=selectedCategory.id,
+                           user_id=login_session['user_id'])
             session.add(newItem)
             session.commit()
-        return redirect(url_for('showItems', cat_name = selectedCategory.name))
+        return redirect(url_for('showItems', cat_name=selectedCategory.name))
     else:
-        return render_template('newitem.html', catalog = catalog)
+        return render_template('newitem.html', catalog=catalog)
 
-#Edit Item
-@app.route('/catalog/<string:item_name>/<int:category_id>/edit/', methods=['GET', 'POST'])
+
+# Edit Item
+@app.route('/catalog/<string:item_name>/<int:category_id>/edit/',
+           methods=['GET', 'POST'])
 def editItem(item_name, category_id):
     catalog = session.query(Category).order_by(asc(Category.name))
-    itemToEdit = session.query(Item).filter_by(name = item_name, category_id = category_id).one()
-    selectedCategory = session.query(Category).filter_by(id = category_id).one()
+    itemToEdit = session.query(Item).filter_by(name=item_name,
+                                               category_id=category_id).one()
+    selectedCategory = session.query(Category).filter_by(id=category_id).one()
+    if itemToEdit.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized \
+            to edit this item. Please create your own item in order to \
+            edit.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         if request.form['title']:
             itemToEdit.name = request.form['title']
@@ -243,41 +274,58 @@ def editItem(item_name, category_id):
             itemToEdit.category_id = category_id
         session.add(itemToEdit)
         session.commit()
-        return redirect(url_for('showOneItem', cat_name = selectedCategory.name, item_name = itemToEdit.name))
+        return redirect(url_for('showOneItem',
+                        cat_name=selectedCategory.name,
+                        item_name=itemToEdit.name))
     else:
-        return render_template('edititem.html', catalog = catalog, item = itemToEdit)
+        return render_template('edititem.html',
+                               catalog=catalog, item=itemToEdit)
 
-#Delete Item
-@app.route('/catalog/<string:item_name>/<int:category_id>/delete', methods=['GET', 'POST'])
+
+# Delete Item
+@app.route('/catalog/<string:item_name>/<int:category_id>/delete',
+           methods=['GET', 'POST'])
 def deleteItem(item_name, category_id):
-    itemToDelete = session.query(Item).filter_by(name = item_name, category_id = category_id).one()
-    selectedCategory = session.query(Category).filter_by(id = category_id).one()
+    itemToDelete = session.query(Item).filter_by(name=item_name,
+                                                 category_id=category_id).one()
+    selectedCategory = session.query(Category).\
+        filter_by(id=category_id).one()
+    if itemToDelete.user_id != login_session['user_id']:
+        return "<script>function myFunction() {alert('You are not authorized \
+            to delete this item. Please create your own item in order \
+            to delete.');}</script><body onload='myFunction()'>"
     if request.method == 'POST':
         session.delete(itemToDelete)
         session.commit()
-        return redirect(url_for('showItems', cat_name = selectedCategory.name))
+        return redirect(url_for('showItems',
+                        cat_name=selectedCategory.name))
     else:
-        return render_template('deleteitem.html', item = itemToDelete)
+        return render_template('deleteitem.html', item=itemToDelete)
 
-#auxiliary functions for User login control
+
+# auxiliary functions for User login control
 def createUser(login_session):
-    newUser = User(name = login_session['username'], email =
-        login_session['email'], picture = login_session['picture'])
+    newUser = User(name=login_session['username'],
+                   email=login_session['email'],
+                   picture=login_session['picture'])
     session.add(newUser)
     session.commit()
-    user = session.query(User).filter_by(email = login_session['email']).one()
+    user = session.query(User).filter_by(email=login_session['email']).one()
     return user.id
 
+
 def getUserInfo(user_id):
-    user = session.query(User).filter_by(id = user_id).one()
+    user = session.query(User).filter_by(id=user_id).one()
     return user
+
 
 def getUserID(email):
     try:
-        user = session.query(User).filter_by(email = email).one()
+        user = session.query(User).filter_by(email=email).one()
         return user.id
     except:
         return None
+
 
 def get_count(q):
     count_q = q.statement.with_only_columns([func.count()]).order_by(None)
